+++ Backwards/src/Types/FunctionValue.cpp
@@ -31,7 +31,7 @@ OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
 #include "Backwards/Types/ValueType.h"
 
-#include "Backwards/Types/DoubleValue.h"
+#include "Backwards/Types/FloatValue.h"
 #include "Backwards/Types/StringValue.h"
 #include "Backwards/Types/ArrayValue.h"
 #include "Backwards/Types/DictionaryValue.h"
@@ -71,7 +71,7 @@ namespace Types
 
    IMPLEMENTVISITOR(FunctionValue)
 
-   bool FunctionValue::sort (const DoubleValue&) const
+   bool FunctionValue::sort (const FloatValue&) const
     {
       return true;
     }
